requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0
pydantic>=2.0.0
chromadb>=0.5.0
python-dotenv>=1.0.0
wordcloud>=1.9.0
matplotlib>=3.7.0
jieba>=0.42.1
sentence-transformers>=2.2.2
sentencepiece>=0.1.99
streamlit>=1.30.0
